Separate album partial, CSS for emails.
Gmail doesn’t support grid or flexbox, so we need to offer a more old-school approach to displaying albums. This _should_ fix the issues in #6.

And given we have a separate album partial, it means we can avoid the ‘purchased’ button - that state wasn’t being hidden in some mail clients, and it’s pointless from the context of the email anyway.