Display basic external purchasing links.
The styling could of course be better, but it’s a start. Also, I’ve reworked things so the album component isn’t a full link - given there may be a variety of purchase locations, this old approach wouldn’t make sense.

Also: opting just for Bandcamp and (generic) iTunes at the moment.